+++ irelia/src/utils/mod.rs
@@ -1,7 +1,5 @@
 #[cfg(any(feature = "ws", feature = "rest"))]
-mod encoder;
-#[cfg(any(feature = "ws", feature = "rest"))]
 pub(crate) mod process_info;
 #[cfg(any(feature = "in_game", feature = "rest"))]
-pub(crate) mod request;
+pub(crate) mod requests;
 pub(crate) mod setup_tls;